requirements.txt
# Core Django
Django>=5.1.0
django-cors-headers>=4.3.0
python-decouple>=3.8

# Social Authentication
django-allauth>=0.61.0

# Database drivers
psycopg2-binary>=2.9.9  # PostgreSQL
mysqlclient>=2.2.0      # MySQL
# For SQLite, no additional package needed

# Django REST Framework (for API)
djangorestframework>=3.14.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0

# AI/ML Libraries
tensorflow>=2.15.0  # or torch>=2.1.0 for PyTorch
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Forms and validation
django-crispy-forms>=2.1
crispy-bootstrap5>=2.0.0

# File uploads
django-cleanup>=8.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Security
django-environ>=0.11.2

# Testing
coverage>=7.4.0
factory-boy>=3.3.0

# Development
django-debug-toolbar>=4.2.0
